Recommended profile should contain all rules from default profile
