+++ sonarts-sq-plugin/sonar-typescript-plugin/src/main/resources/org/sonar/l10n/typescript/rules/typescript/Sonar_way_recommended_profile.json
@@ -43,6 +43,7 @@
     "S1862",
     "S1871",
     "S1874",
+    "S1940",
     "S2068",
     "S2123",
     "S2201",
@@ -55,6 +56,7 @@
     "S2757",
     "S2814",
     "S2870",
+    "S2871",
     "S2933",
     "S2966",
     "S3257",
@@ -62,6 +64,7 @@
     "S3498",
     "S3504",
     "S3512",
+    "S3516",
     "S3533",
     "S3616",
     "S3626",
@@ -91,6 +94,7 @@
     "S4144",
     "S4156",
     "S4157",
+    "S4158",
     "S4275",
     "S4322",
     "S4323",
